feat: Add permission check for send command

+++ src/cmds.py
@@ -1,4 +1,5 @@
 from logger import logger
+import config
 
 
 class Commands:
@@ -12,5 +13,10 @@ class Commands:
 
         @bot.command()
         async def send(ctx, *, message: str):
+            if not any(role.id in config.ADMIN_ROLES_IDS for role in ctx.author.roles):
+                logger.warning(
+                    f"El usuario '{ctx.author}' intentó enviar un mensaje sin los permisos necesarios."
+                )
+                return
             await ctx.message.delete()
             await ctx.send(message)
